extracting PropagationRegistry to separate class and including ExceptionCatching decorators inside for narrower visibility
